+++ GRDB/Core/FetchRequest.swift
@@ -107,6 +107,28 @@ extension FetchRequest {
 
 // MARK: - PreparedRequest
 
+/// A closure executed before a supplementary fetch is performed.
+///
+/// Support for `Database.dumpRequest`.
+///
+/// - parameter request: The supplementary request
+/// - parameter keyPath: The key path target of the supplementary fetch.
+typealias WillExecuteSupplementaryRequest = (_ request: AnyFetchRequest<Row>, _ keyPath: [String]) throws -> Void
+
+/// A closure that performs supplementary fetches.
+///
+/// Support for eager loading of hasMany associations.
+///
+/// - parameter db: A database connection.
+/// - parameter rows: The rows that are modified by the supplementary fetch.
+/// - parameter willExecuteSupplementaryRequest: A closure to execute before
+///   performing supplementary fetches.
+typealias SupplementaryFetch = (
+    _ db: Database,
+    _ rows: [Row],
+    _ willExecuteSupplementaryRequest: WillExecuteSupplementaryRequest?)
+throws -> Void
+
 /// A `PreparedRequest` is a request that is ready to be executed.
 public struct PreparedRequest {
     /// A prepared statement with bound parameters.
@@ -115,13 +137,14 @@ public struct PreparedRequest {
     /// An eventual adapter for rows fetched by the select statement.
     public var adapter: (any RowAdapter)?
     
+    /// A closure that performs supplementary fetches.
     /// Support for eager loading of hasMany associations.
-    var supplementaryFetch: ((Database, [Row]) throws -> Void)?
+    var supplementaryFetch: SupplementaryFetch?
     
     init(
         statement: Statement,
         adapter: (any RowAdapter)?,
-        supplementaryFetch: ((Database, [Row]) throws -> Void)? = nil)
+        supplementaryFetch: SupplementaryFetch? = nil)
     {
         self.statement = statement
         self.adapter = adapter